chore: remove gulp dependency from project
